// Frontend/Angular-NVT/src/app/locations/model/sector.ts
export class Sector {
    id: number;
    title: string;
    type: string;
    // tslint:disable-next-line: variable-name
    max_guests: number;
    // tslint:disable-next-line: variable-name
    row_num: number;
    // tslint:disable-next-line: variable-name
    column_num: number;
    left: number;
    top: number;
    height: number;
    width: number;
    angle: number;
}